storage: extract replica gossip helpers
Release note: None
